Add the useOverlay hook
