--- lib/components/OverlayTrigger/index.ts
@@ -1,1 +1,0 @@
-export {OverlayTrigger} from "./OverlayTrigger"
