+++ lib/components/Suggestions/Suggestions.tsx
@@ -2,8 +2,10 @@ import {ForwardedRef, forwardRef, useMemo, useState} from "react";
 import {OverlayProps} from "../../types";
 import {KEY} from "../../constants";
 import {useDownOf} from "../../utils/useDownOf";
+import {useOverlay} from "../../utils/useOverlay";
 
-export const Suggestions = forwardRef(({trigger, style, onSelect}: OverlayProps, ref: ForwardedRef<any>) => {
+export const Suggestions = () => {
+    const {trigger, onSelect, style, ref} = useOverlay()
     const [active, setActive] = useState(NaN)
     const filtered = useMemo(
         () => trigger.option.data.filter(s => s.toLowerCase().indexOf(trigger.value.toLowerCase()) > -1),
@@ -45,4 +47,4 @@ export const Suggestions = forwardRef(({trigger, style, onSelect}: OverlayProps,
             })}
         </ul>
     )
-})
+}
