+++ lib/components/Whisper.tsx
@@ -1,12 +1,12 @@
-import {OverlayTrigger} from "./OverlayTrigger";
 import {memo} from "react";
 import {useSelector} from "../utils/useSelector";
+import {Suggestions} from "./Suggestions";
 
 export const Whisper = memo(() => {
     const trigger = useSelector(state => state.trigger)
+    const Overlay = useSelector(state => state.Overlay ?? Suggestions)
 
-    if (trigger) return <OverlayTrigger key={trigger.option.index} {...trigger}/>
-    return null
+    return trigger ? <Overlay key={trigger.option.index}/> : null;
 })
 
 Whisper.displayName = 'Whisper'
