+++ lib/index.ts
@@ -3,3 +3,4 @@ export {Option} from "./components/Option"
 export {annotate, denote} from "./utils"
 export {createMarkedInput} from "./utils/createMarkedInput"
 export {useMark} from "./utils/useMark";
+export {useOverlay} from "./utils/useOverlay";
